Update BubbleSort.py
— What is Bubble Sort?

Bubblesort, sometimes called sinking sort, is a simple
sorting algorithm. It is a simple algorithm, but inefficient. 
It is often used as an illustration of an algorithm because
of its simplicity and because it is easy to explain. 

— How does Bubble Sort work?

Start from the leftmost element, compare it to the neighboring
element to the right, and put them in the correct order. 
— When should I use this sorting algorithm?

Honestly? Never. With an average complexity of О(n²), it's incredibly
slow, especially when handling a lot of data.